src/transactions/pyproject.toml
[project]
name = "transactions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=3.4.3",
    "fastapi>=0.115.12",
    "pydantic>=2.11.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
    "web3>=7.10.0",
]
